# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
    "title": "Challenge Chapter 5",
    "contact": {},
    "version": "1.0"
    },
    "servers": [
    {
        "url": "http://localhost:3000/api/v1",
        "variables": {}
    }
    ],
    "paths": {
...
    "/users/{id}": {
        "get": {
        "tags": [
            "Shops"
        ],
        "summary": "Get User By Id",
        "operationId": "GetUserById",
        "parameters": [
            {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
                "type": "integer",
                "format": "int32",
                "example": 3
            }
            }
        ],
        "responses": {
            "200": {
            "description": "",
            "headers": {}
            }
        },
        "deprecated": false
        }
    },
...
    "/users/delete/{id}": {
        "delete": {
        "tags": [
            "Shops"
        ],
        "summary": "Delete user",
        "operationId": "DeleteUser",
        "parameters": [
            {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
                "type": "integer",
                "format": "int32",
                "example": 4
            }
            }
        ],
        "responses": {
            "200": {
            "description": "",
            "headers": {}
            }
        },
        "deprecated": false,
        "security": [
            {
            "bearer": []
            }
        ]
        }
    },
...
    },
    "components": {
...
    "securitySchemes": {
        "bearer": {
        "type": "http",
        "scheme": "bearer"
        }
    }
    },
    "security": [],
...
}